requirements.txt
django
django-cors-headers
djangorestframework
sentence-transformers
numpy<2
transformers
datasets
torch